Mantenedor de Pais
Se hizo la barra lateral y se creó el mantenedor de pais

+++ desarrollo-ucsc/database/migrations/2025_04_18_221524_create_region_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->integer('id_region');
             $table->primary('id_region');
             
-            $table->integer('id_pais');
+            $table->unsignedBigInteger('id_pais');
             $table->foreign('id_pais')->references('id_pais')->on('pais');
             
             $table->string('nombre_region');
